requirements.txt
Django<1.10
django-bootstrap3
djangorestframework
markdown
django-filter
django-crispy-forms
Pillow